(PC-25661) feat(offer): extract offer body from offer content
