+++ src/features/offer/components/OfferContent/OfferContent.tsx
@@ -1,43 +1,27 @@
 import { useNavigation } from '@react-navigation/native'
 import React, { FunctionComponent, useCallback, useEffect } from 'react'
-import { NativeScrollEvent, NativeSyntheticEvent, Platform, View } from 'react-native'
+import { NativeScrollEvent, NativeSyntheticEvent, Platform } from 'react-native'
 import { IOScrollView as IntersectionObserverScrollView } from 'react-native-intersection-observer'
-import styled, { useTheme } from 'styled-components/native'
+import styled from 'styled-components/native'
 
-import { OfferResponse, SearchGroupResponseModelv2, SubcategoryIdEnum } from 'api/gen'
+import { OfferResponse, SearchGroupResponseModelv2 } from 'api/gen'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
-import { MovieScreeningCalendar } from 'features/offer/components/MovieScreeningCalendar/MovieScreeningCalendar'
-import { OfferAbout } from 'features/offer/components/OfferAbout/OfferAbout'
-import { OfferArtists } from 'features/offer/components/OfferArtists/OfferArtists'
+import { OfferBody } from 'features/offer/components/OfferBody/OfferBody'
 import { OfferCTAButton } from 'features/offer/components/OfferCTAButton/OfferCTAButton'
 import { OfferHeader } from 'features/offer/components/OfferHeader/OfferHeader'
-import { OfferMessagingApps } from 'features/offer/components/OfferMessagingApps/OfferMessagingApps'
-import { OfferPlace } from 'features/offer/components/OfferPlace/OfferPlace'
 import { OfferPlaylistList } from 'features/offer/components/OfferPlaylistList/OfferPlaylistList'
-import { OfferPrice } from 'features/offer/components/OfferPrice/OfferPrice'
-import { OfferSummaryInfoList } from 'features/offer/components/OfferSummaryInfoList/OfferSummaryInfoList'
-import { OfferTitle } from 'features/offer/components/OfferTitle/OfferTitle'
-import { OfferVenueButton } from 'features/offer/components/OfferVenueButton/OfferVenueButton'
 import { OfferWebMetaHeader } from 'features/offer/components/OfferWebMetaHeader'
-import { getOfferArtists } from 'features/offer/helpers/getOfferArtists/getOfferArtists'
-import { getOfferPrices } from 'features/offer/helpers/getOfferPrice/getOfferPrice'
-import { getOfferTags } from 'features/offer/helpers/getOfferTags/getOfferTags'
 import { useOfferBatchTracking } from 'features/offer/helpers/useOfferBatchTracking/useOfferBatchTracking'
 import { useOfferPlaylist } from 'features/offer/helpers/useOfferPlaylist/useOfferPlaylist'
-import { useOfferSummaryInfoList } from 'features/offer/helpers/useOfferSummaryInfoList/useOfferSummaryInfoList'
 import { analytics, isCloseToBottom } from 'libs/analytics'
 import { useFeatureFlag } from 'libs/firebase/firestore/featureFlags/useFeatureFlag'
 import { RemoteStoreFeatureFlags } from 'libs/firebase/firestore/types'
 import { useFunctionOnce } from 'libs/hooks'
 import { useLocation } from 'libs/location'
 import { Subcategory } from 'libs/subcategories/types'
-import { FeatureFlag } from 'shared/FeatureFlag/FeatureFlag'
 import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition'
 import { Hero } from 'ui/components/hero/Hero'
-import { SectionWithDivider } from 'ui/components/SectionWithDivider'
-import { Separator } from 'ui/components/Separator'
-import { InformationTags } from 'ui/InformationTags/InformationTags'
-import { getSpacing, Spacer } from 'ui/theme'
+import { Spacer } from 'ui/theme'
 
 type Props = {
   offer: OfferResponse
@@ -53,16 +37,6 @@ export const OfferContent: FunctionComponent<Props> = ({ offer, searchGroupList,
   const { userLocation } = useLocation()
   const { navigate } = useNavigation<UseNavigationType>()
   const enableOfferPreview = useFeatureFlag(RemoteStoreFeatureFlags.WIP_OFFER_PREVIEW)
-  const { isDesktopViewport } = useTheme()
-
-  const extraData = offer.extraData ?? undefined
-  const tags = getOfferTags(subcategory.appLabel, extraData)
-  const artists = getOfferArtists(subcategory.categoryId, offer)
-  const prices = getOfferPrices(offer.stocks)
-
-  const isOfferAMovieScreening = offer.subcategoryId === SubcategoryIdEnum.SEANCE_CINE
-
-  const { summaryInfoItems } = useOfferSummaryInfoList({ offer })
 
   const {
     sameArtistPlaylist,
@@ -135,59 +109,7 @@ export const OfferContent: FunctionComponent<Props> = ({ offer, searchGroupList,
           onPress={onPress}
         />
         <Spacer.Column numberOfSpaces={8} />
-        <InfoContainer>
-          <GroupWithoutGap>
-            <InformationTags tags={tags} />
-            <Spacer.Column numberOfSpaces={4} />
-            <OfferTitle offerName={offer.name} />
-
-            {artists ? (
-              <React.Fragment>
-                <Spacer.Column numberOfSpaces={2} />
-                <OfferArtists artists={artists} />
-              </React.Fragment>
-            ) : null}
-          </GroupWithoutGap>
-
-          {prices ? <OfferPrice prices={prices} /> : null}
-
-          {!offer.venue.isPermanent && summaryInfoItems.length === 0 ? null : (
-            <GroupWithoutGap>
-              {offer.venue.isPermanent ? <OfferVenueButton venue={offer.venue} /> : null}
-
-              {!offer.venue.isPermanent && summaryInfoItems.length === 0 ? null : (
-                <Separator.Horizontal testID="topSeparator" />
-              )}
-
-              {summaryInfoItems.length === 0 ? null : (
-                <OfferSummaryInfoList summaryInfoItems={summaryInfoItems} />
-              )}
-            </GroupWithoutGap>
-          )}
-
-          <OfferAbout offer={offer} />
-        </InfoContainer>
-
-        <OfferPlace offer={offer} isEvent={subcategory.isEvent} />
-        <Spacer.Column numberOfSpaces={8} />
-
-        {isOfferAMovieScreening ? (
-          <FeatureFlag featureFlag={RemoteStoreFeatureFlags.WIP_ENABLE_NEW_XP_CINE_FROM_OFFER}>
-            <MovieScreeningCalendar offer={offer} subcategory={subcategory} />
-          </FeatureFlag>
-        ) : null}
-
-        {isDesktopViewport ? (
-          <ContainerWithoutDivider testID="messagingApp-container-without-divider">
-            <Spacer.Column numberOfSpaces={2} />
-            <OfferMessagingApps offer={offer} />
-          </ContainerWithoutDivider>
-        ) : (
-          <SectionWithDivider visible margin testID="messagingApp-container-with-divider">
-            <Spacer.Column numberOfSpaces={2} />
-            <OfferMessagingApps offer={offer} />
-          </SectionWithDivider>
-        )}
+        <OfferBody offer={offer} subcategory={subcategory} />
 
         <OfferPlaylistList
           offer={offer}
@@ -220,14 +142,3 @@ const Container = styled.View({
 })
 
 const ScrollViewContainer = styled(IntersectionObserverScrollView)({ overflow: 'visible' })
-
-const InfoContainer = styled.View({
-  marginHorizontal: getSpacing(6),
-  gap: getSpacing(6),
-})
-
-const ContainerWithoutDivider = styled.View(({ theme }) => ({
-  marginHorizontal: theme.contentPage.marginHorizontal,
-}))
-
-const GroupWithoutGap = View
