+++ src/features/offer/todo.md
@@ -1,0 +1,20 @@
+# TODO
+
+## Links
+
+- [PC-25661](https://passculture.atlassian.net/browse/PC-25661)
+- [MobTime](https://mobtime.hadrienmp.fr/mob/decouverte)
+
+---
+
+## Tasks
+
+- [ ] Créer un composant avec le contenu de la page offre des tags au bloc des réseaux sociaux
+- [ ] Afficher correctement l'image suivant qu'on est en affichage desktop ou mobile
+- [ ] Utiliser le composant BookingButton pour l'affichage desktop
+
+---
+
+## Tasks for another US
+
+- [ ]
